Created functioning basic app
Added html templates into the app to provide visualisations,
these are copied from the 2nd project in the nd025 course but
alterations have been made. I have included an upload_image.html
which I need to alter to provide better visualisations of predictions.

I have created the define_model.py script in the model folder to define,
train, evaluate and save a neural network off the back of VGG19 model. THis
created the other hdf5 files. I've also included a prediction.py file which
defines all the functions necessary for predicting a classification of dog breed.

I've added functionality in the run.py file to add visualisations to the app.

The samples folder is included as a folder which images get uploaded to.

The dog_app notebook has been fully run and outlines the initial exploration.

TO DO:
- Update upload_image.html
- Add more visualisations to run.py
- Create a new notebok for exploratory analysis
- Add comments and function doc strings
- Update README.md with direction on how to run, problem statement, etc.